+++ src/model.ts
@@ -134,7 +134,6 @@ abstract class $Node {
     const extractChildren = (owner: any): List<Node> => {
       if (isNode(owner)) return [owner]
       if (isArray(owner)) return owner.flatMap(extractChildren)
-      if (owner instanceof Object) return values(owner).flatMap(extractChildren) // TODO: Remove once we drop constructors
       return []
     }
 
@@ -182,10 +181,8 @@ abstract class $Node {
 
   transform(tx: (node: Node) => Node): this {
     const applyTransform = (value: any): any => {
-      if (typeof value === 'function') return value
       if (isArray(value)) return value.map(applyTransform)
       if (isNode(value)) return value.copy(mapObject(applyTransform, tx(value as any)))
-      if (value instanceof Object) return mapObject(applyTransform, value) // TODO: Remove once we drop constructors
       return value
     }