+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## v3.0.0
-- Removed Constructors
-- Removed instantiation with unnamed arguments
+- Implemented Wollok 3.0.0 support
+    - Removed Constructors
+    - Removed instantiation with unnamed arguments
+    - Removed Fixtures
 
 
 ## v2.3.0